Compile typescript in Github
Previously I'd compile the typescript manually, check the compiled
javascript to the repo, and push to origin. Now reduce the possibility
of divergence between the checked in JS and TS code, I add a github
workflow to compile the typescript and check-in to a separate gh-page
branch. The Github Page takes from this gh-page branch rather than from
the mainline, and mainline is now free of compiled code.

__Testing done__

1. `npm run deploy` succeeds

(To see if the workflow works)